+++ src/collections.js
@@ -5,7 +5,7 @@ var deepEqual = require('./deep-equal');
 var engine = {};
 
 
-// b is assumed to have one element and it tests whether b[0] is in a
+// "b" is assumed to have one element and it tests whether "b[0]" is in "a"
 function containsImpl(a,b){
   for(var i = 0; i < a.length; i++){
     if(deepEqual(a[i], b[0])) { return true; }